Command Line functionality in uploadfiles.php

+++ cgi-bin/upload_files.php
@@ -3,16 +3,33 @@
 	require 'get_labels.php';
 	require 'get_rawfiles_names.php';
 	
-
+	
 	
 	// upload_files.php does not only upload the files to the server but at the same time finds the rawfiles contained in the uploaded file and
 	// the labels (if any) that are contained in the file using get_rawfiles_names() and get_labels() from the corresponding php files that are loaded above.
 	// upload_files also guesses the experiment type - that is if the experiment was Metabolically labelled (L), Isobarically labelled (IL) or Label Free (LF)
-	//
+	
+	// upload_files can be run using command line, it accepts the following arguments:
+	// the first arg should always be "CL" standing for command line
+	// the second contains the path of the file to read
+	
+	$CL_run = false;
+	if (isset($argv[1]) && $argv[1] == "CL")
+	{
+		print("Upload files is run using Command Line\n");
+		$_POST["server_side"] = "true";
+		$_POST['thefile'] = $argv[2];
+		$_POST["session_id"] = 1;
+		$CL_run = true;
+		$file_copied_successfully = true;
+	}
+	
+	
 	// PS assumes that some patterns in headers lead to a specific conclusion. For example if the headers contain "Spectrum File" it is assumed that this
 	// file is a PD file since only PD files contain this kind of  (the uploaded file could be: MQ for MQ_evidence files, MQP for MQ_proteinGroups files and PD for PD psm files)
 	//
 	// This table shows which main patterns lead to which conlusion:
+	
 	//
 	// +------------------------------+------------------------------------+
 	// |            Pattern           |             Conclusion             |
@@ -49,14 +66,14 @@
 	// For more information about _get_labels_aguments_sets and unique_patterns see the comments below
 	
 	$_get_labels_aguments_sets = array(
-		'PD' => array(
-			'L' => array(
-				array('/^([^\s]+)\/([^\s]+)$/i', '/Modifications/i', '/\((?:[^:]+?:)?(.+?)\)(;|$)/'), //Proteome Discoverer
-				array('/Abundance: (\d+)/i', '/Modifications/i', '/\((?:[^:]+?:)?(.+?)\)(;|$)/'), //Proteome Discoverer
-			)
-		),
-		'MQ' => array(
-			'L' => array(
+	'PD' => array(
+		'L' => array(
+			array('/^([^\s]+)\/([^\s]+)$/i', '/Modifications/i', '/\((?:[^:]+?:)?(.+?)\)(;|$)/'), //Proteome Discoverer
+			array('/Abundance: (\d+)/i', '/Modifications/i', '/\((?:[^:]+?:)?(.+?)\)(;|$)/'), //Proteome Discoverer
+		)
+	),
+	'MQ' => array(
+		'L' => array(
 				array('/^Ratio ([^\s]+)\/([^\s]+)/i', null, null), //MaxQuant
 				array('/^Reporter intensity ([0-9]+)/i', null, null), //MaxQuant, MS/MS-multiplexing (reporter ions), e.g. iTRAQ
 			)
@@ -64,15 +81,15 @@
 	);
 	
 	$unique_patterns = array(
-		'PD' => array(
-			'L' => array('/Quan Channel/i'),
-			'LF' => array(null), // there is no unique header for PD-LF
-			'IL' => array('/\d+\/\d+/', 'Abundance: \d+')
+			'PD' => array(
+				'L' => array('/Quan Channel/i'),
+				'LF' => array(null), // there is no unique header for PD-LF
+				'IL' => array('/\d+\/\d+/', 'Abundance: \d+')
 		),
-		'MQ' => array(
-			'L' => array('/Labeling State/i'),
-			'LF' => array(null), // there is no unique header for MQ-LF
-			'IL' => array('/Reporter intensity/i')
+			'MQ' => array(
+				'L' => array('/Labeling State/i'),
+				'LF' => array(null), // there is no unique header for MQ-LF
+				'IL' => array('/Reporter intensity/i')
 		)
 	);
 	
@@ -103,48 +120,59 @@
 	if (isset($name)) {
 		if (!empty($name)) {
 			
-			// Create a new session directory
-			$location = $document_root . '/uploads/' . $_POST["session_id"];
-			if (!file_exists($location) && !is_dir($location)) {
-				if (!mkdir($location, 0777, true)) {
-					$server_response['mkdir_msg'] = "The directory $location already exists, probably created from another upload_files instance.";
-					//Sometimes upload files tries to create an existing directory even if file_exists($location) returns true, check again if the file exists and display the error if so
-					if (!file_exists($location) && !is_dir($location)) {
-						$server_response['msg'] = "The directory $location could not be created ('mkdir' returned FALSE).";
+			if (!$CL_run)
+			{
+				// Create a new session directory
+				$location = $document_root . '/uploads/' . $_POST["session_id"];
+				if (!file_exists($location) && !is_dir($location)) {
+					if (!mkdir($location, 0777, true)) {
+						$server_response['mkdir_msg'] = "The directory $location already exists, probably created from another upload_files instance.";
+						//Sometimes upload files tries to create an existing directory even if file_exists($location) returns true, check again if the file exists and display the error if so
+						if (!file_exists($location) && !is_dir($location)) {
+							$server_response['msg'] = "The directory $location could not be created ('mkdir' returned FALSE).";
+							goto end;
+						}
+					}
+				}
+				
+				// Copy the file to the session folder: if the file is server sided
+				// simply copy it from test data folder. If ot was manually uploaded
+				// from the user, apache server uses a temp_name to store it and move_uploaded_file is the
+				// dedicated php function to copy it to another folder - here the session folder
+				$file_copied_successfully = false;
+				if ($server_side_file)
+				{
+					$file_copied_successfully = copy($document_root . "/test data/" . $tmp_name, $location . '/' . $name);
+				}
+				else
+				{
+					if (!file_exists($location . "/" . $name))
+					{
+						$file_copied_successfully = move_uploaded_file($tmp_name, $location . "/" . $name);
+					}
+					else
+					{
+						$server_response['success'] = false;
+						$server_response['msg'] = "A file with the same name ($name) has already been uploaded, try again later";
 						goto end;
 					}
 				}
 			}
-
-			// Copy the file to the session folder: if the file is server sided
-			// simply copy it from test data folder. If ot was manually uploaded
-			// from the user, apache server uses a temp_name to store it and move_uploaded_file is the
-			// dedicated php function to copy it to another folder - here the session folder
-			$file_copied_successfully = false;
-			if ($server_side_file)
+			
+			if (!$CL_run)
 			{
-				$file_copied_successfully = copy($document_root . "/test data/" . $tmp_name, $location . '/' . $name);
+				$full_path = $location . '/' . $name;
 			}
 			else
 			{
-				if (!file_exists($location . "/" . $name))
-				{
-					$file_copied_successfully = move_uploaded_file($tmp_name, $location . "/" . $name);
-				}
-				else
-				{
-					$server_response['success'] = false;
-					$server_response['msg'] = "A file with the same name ($name) has already been uploaded, try again later";
-					goto end;
-				}
+				$full_path = $name;
 			}
 			
-			
 			$handle = null;
-			if ($file_copied_successfully && ($handle = fopen($location . '/' . $name, "r")))
+			if ($file_copied_successfully && ($handle = fopen($full_path, "r")))
 			{
-		
-		
+				
+				
 				// First decide if the processing program is PD or MQ and thye filetype (MQ, MQP or PD)
 				// Get the headers of the file
 				$first_line = fgetcsv($handle, 0, "\t");
@@ -225,12 +253,12 @@
 					/*
 						Array
 						(
-							[0] => Array
-								(
-									[0] => /^([^\s]+)\/([^\s]+)$/i               (label header pattern)
-									[1] => /Modifications/i                      (label definition header pattern)
-									[2] => /\((?:[^:]+?:)?(.+?)\)(;|$)/          (label definition pattern)
-								)
+						[0] => Array
+						(
+						[0] => /^([^\s]+)\/([^\s]+)$/i               (label header pattern)
+						[1] => /Modifications/i                      (label definition header pattern)
+						[2] => /\((?:[^:]+?:)?(.+?)\)(;|$)/          (label definition pattern)
+						)
 						)
 					*/
 					
@@ -254,21 +282,21 @@
 						/*
 							Array
 							(
-								[0] => /^([^\s]+)\/([^\s]+)$/i               (label header pattern)
-								[1] => /Modifications/i                      (label definition header pattern)
-								[2] => /\((?:[^:]+?:)?(.+?)\)(;|$)/          (label definition pattern)
+							[0] => /^([^\s]+)\/([^\s]+)$/i               (label header pattern)
+							[1] => /Modifications/i                      (label definition header pattern)
+							[2] => /\((?:[^:]+?:)?(.+?)\)(;|$)/          (label definition pattern)
 							)
 						*/
 						
-						$tmp = get_labels($location . '/' . $name, $argset[0], $argset[1], $argset[2]);
+						$tmp = get_labels($full_path, $argset[0], $argset[1], $argset[2]);
 						
 						// tmp is an array of two arrays.
 						// tmp[0] contains all labels detected in the file e.g.
 						/*
 							Array
 							(
-								[0] => Medium
-								[1] => Light
+							[0] => Medium
+							[1] => Light
 							)
 						*/
 						// and tmp[1] is null
@@ -277,35 +305,35 @@
 						/*
 							if (count($tmp[0]) > 0)
 							{
-								// If get_labels returned at least one label definition
-								if (count($tmp[1]) > 0)
-								{
-									$okdefs = 0;
-									// For each label definition:
-									foreach ($tmp[1] as $lbldef)
-									{
-										// Make sure that the label definition does not contain space characters,
-										// commas, semicolons and colons
-										if (preg_match('/[\s,;\:]/i', $lbldef) == 0)
-										{
-											$okdefs++;
-										}
-										else
-										{
-											array_push($server_response['skipped_labels'], $lbldef);
-										}
-									}
-									// If all definitions are OK stop calling get_labels
-									if ($okdefs == count($tmp[1]))
-									{
-										break;
-									}
-								}
-								else
-								{
-									// If no definitions were returned stop calling get_labels
-									break;
-								}
+							// If get_labels returned at least one label definition
+							if (count($tmp[1]) > 0)
+							{
+							$okdefs = 0;
+							// For each label definition:
+							foreach ($tmp[1] as $lbldef)
+							{
+							// Make sure that the label definition does not contain space characters,
+							// commas, semicolons and colons
+							if (preg_match('/[\s,;\:]/i', $lbldef) == 0)
+							{
+							$okdefs++;
+							}
+							else
+							{
+							array_push($server_response['skipped_labels'], $lbldef);
+							}
+							}
+							// If all definitions are OK stop calling get_labels
+							if ($okdefs == count($tmp[1]))
+							{
+							break;
+							}
+							}
+							else
+							{
+							// If no definitions were returned stop calling get_labels
+							break;
+							}
 							}
 						*/
 						// If we found labels just break testing patterns
@@ -323,25 +351,34 @@
 					
 					// In a similar manner, get_rawfiles_names gets all raw_files contained in the file: its definition is almost identical to this of get_labels
 					
-					$server_response['raw_filesnames'] = get_rawfiles_names($location . '/' . $name, '/file/i');
+					$server_response['raw_filesnames'] = get_rawfiles_names($full_path, '/file/i');
 					
 					
-					if (count($server_response['raw_filesnames']) == 0)
+					if (count($server_response['raw_filesnames']) == 0 && !$CL_run)
 					{
 						error_log("[client: " . $_SERVER['REMOTE_ADDR'] . "] Could not retrieve replicate information (raw files names) from data file " . $name);
 					}
 					// Rename the files to msdiffexp_peptide.txt fro PD and MQ files and to msdiffexp_protein.txt for MQP files
-					rename($location . '/' . $name, $location . '/msdiffexp_peptide.txt');
+					if (!$CL_run)
+					{
+						rename($location . '/' . $name, $location . '/msdiffexp_peptide.txt');
+					}
 				}
 				elseif ($dtype == 'MQP')
 				{
 					$server_response['file_type'] = $dtype;
-					rename($location . '/' . $name, $location . '/msdiffexp_protein.txt');
+					if (!$CL_run)
+					{
+						rename($location . '/' . $name, $location . '/msdiffexp_protein.txt');
+					}
 				}
 				else
 				{
 					$server_response['file_type'] = "unknown";
-					unlink($location . '/' . $name);
+					if (!$CL_run)
+					{
+						unlink($location . '/' . $name);
+					}
 					$server_response['success'] = true;
 					$server_response['msg'] = "The file $name is not valid";
 					goto end;
@@ -364,8 +401,23 @@
 	}
 	
 	end:
-	error_log("[client: " . $_SERVER['REMOTE_ADDR'] . "] upload_files.php [" . $_POST["session_id"] . " " . $name . " ]> Success: " . ($server_response['success'] ? 'Yes' : 'No') . " | Message: " . $server_response['msg']);
-	//Send info back to the client
-	header('Content-type: application/json');
-	echo json_encode($server_response);
+	if (!$CL_run)
+	{
+		error_log("[client: " . $_SERVER['REMOTE_ADDR'] . "] upload_files.php [" . $_POST["session_id"] . " " . $name . " ]> Success: " . ($server_response['success'] ? 'Yes' : 'No') . " | Message: " . $server_response['msg']);
+		//Send info back to the client
+		header('Content-type: application/json');
+		echo json_encode($server_response);
+	}
+	else
+	{
+		// In case the program was run by command line print the necessary results back to the user
+		print("Filetype:\n");
+		print($server_response['file_type']);
+		print("\nExperiment type:\n");
+		print($server_response['exp_type']);
+		print("\n\nPeptide label names from file:\n");
+		print_r($server_response['peptide_labels_names']);
+		print("\n\nRaw files detected:\n");
+		print_r($server_response['raw_filesnames']);
+	}
 ?>
